Move primitive rendering out of DeferredLightingEffect
